Drop ActionCommand and Raw form
`foo="a b"` and `foo=["a", "b"]` are equivalent, as are `foo="a"` and
`foo=["a"]`.  The `Raw` field only existed to remember the difference,
which isn't important.  And having the `Raw` field there made it possible
for the two fields of the struct to get out of sync, or have the `Raw`
field get misinterpreted when a caller should just use the `Parsed` field.

So this commit drops it.  `ActionCommand` is gone entirely.  `Args` and
`Command` are just `[]string`.

Resolves #5